rename data.ts to content.ts
