+++ src/components/code.tsx
@@ -3,7 +3,7 @@ import theme from 'prism-react-renderer/themes/nightOwl'
 import React from 'react'
 import { Button } from 'reactstrap'
 import { CodeProps } from '../types'
-import { content } from '../utils/data'
+import { content } from '../utils/content'
 import { copyToClipboard } from '../utils/copy-to-clipboard'
 
 const Code = ({ codeString, language }: CodeProps) => {
