+++ src/components/contact.tsx
@@ -7,7 +7,7 @@ import NewTabLink from './newTabLink'
 import useSiteMetadata from '../hooks/useSiteMetadata'
 import useSiteSettings from '../hooks/useSiteSettings'
 import useSiteUrls from '../hooks/useSiteUrls'
-import { content } from '../utils/data'
+import { content } from '../utils/content'
 import linkColor from '../utils/linkColor'
 import validateEmail from '../utils/validateEmail'
 import { SectionProps } from '../types'
